+++ NEWS.md
@@ -1,3 +1,13 @@
+# inlmisc 0.3.0.9000
+
+- Add `AddCertificate` function, used to add a X.509 certificate to your CA bundle.
+
+- Make **httr** a suggested package.
+
+- In `RecreateLibrary` function, add `local` argument, file paths of files containing builds of packages.
+
+- In `AddGradientLegend` function, add `n` argument, the desired number of tick-marks to be drawn.
+
 # inlmisc 0.3.0
 
 - Add `Grid2Polygons` function from the **Grid2Polygons** package.